+++ src/types/processDto.ts
@@ -23,10 +23,12 @@ export enum IpcChannels {
 	THEME_CHANGED = 'theme-changed',
 	OPEN_IN_BROWSER = 'open-in-browser',
 	OPEN_PATH = 'open-path',
-	OPEN_SELECT_FOLDER_DIALOG = 'open-select-directory-dialog',
+	OPEN_SELECT_IMAGES_FOLDER_DIALOG = 'open-select-images-directory-dialog',
+	OPEN_SELECT_FOLDER_DIALOG = 'open-select-folder-dialog',
 	OPEN_SELECT_EXPORTED_DATA_FILE_DIALOG = 'open-select-exported-data-file-dialog',
 	SAVE_EXPORTED_DATA = 'save-exported-data',
 	OPEN_IMPORT_DATA_DIALOG = 'open-import-data-dialog',
+	EXPORT_POSTS = 'export-posts',
 }
 
 export interface ExportDataDto {